requirements.txt
streamlit
pinecone-client
langchain-ollama
langchain-community
sentence-transformers
langchain
python-dotenv



# virtual environment
.venv/

# python cache
__pycache__/

# env files
.env

# streamlit
.streamlit/

# executables (IMPORTANT)
*.exe

